practica2/requirements.txt
requests
pycryptodomex
argparse
